+++ Cargo.toml
@@ -2,8 +2,9 @@
 resolver = "2"
 members = [
     "xrpl-wasm-stdlib",
-    "xrpl-wasm-stdlib/xrpl-parameter-macro",
     "xrpl-wasm-stdlib/xrpl-macros",
+    "xrpl-contract-stdlib",
+    "xrpl-contract-stdlib/xrpl-parameter-macro",
 ]
 exclude = [
     "examples",
