+++ e2e-tests/contract_data/Cargo.toml
@@ -23,4 +23,4 @@ overflow-checks = false # Disable overflow checks (use cautiously)
 panic = "unwind"
 
 [dependencies]
-xrpl-wasm-stdlib = { path = "../../xrpl-wasm-stdlib" }
+xrpl-contract-stdlib = { path = "../../xrpl-contract-stdlib" }
