+++ e2e-tests/parameter_macro/Cargo.toml
@@ -26,4 +26,4 @@ overflow-checks = false # Disable overflow checks (use cautiously)
 panic = "unwind"
 
 [dependencies]
-xrpl-wasm-stdlib = { path = "../../xrpl-wasm-stdlib" }
+xrpl-contract-stdlib = { path = "../../xrpl-contract-stdlib" }
