+++ xrpl-contract-stdlib/README.md
@@ -1,0 +1,9 @@
+# xrpl-contract-stdlib
+
+Smart Contract support crate for XRPL WebAssembly contracts. Provides the
+contract-specific entry-point context, parameter extraction, contract data
+persistence, event emission, and transaction-building helpers that smart
+contracts on the XRPL use.
+
+Built on top of `xrpl-wasm-stdlib`, which contains the shared host bindings
+and type system.
