+++ xrpl-wasm-stdlib/Cargo.toml
@@ -24,7 +24,6 @@ crate-type = ["lib"]
 test-host-bindings = ["dep:mockall"]
 
 [dependencies]
-xrpl-parameter-macro = { path = "xrpl-parameter-macro" }
 xrpl-macros = { version = "0.1.0", path = "xrpl-macros" }
 mockall = { version = "0.14.0", optional = true }
