+++ xrpl-wasm-stdlib/src/core/current_tx/traits.rs
@@ -413,16 +413,6 @@ pub trait EscrowFinishFields: TransactionCommonFields {
     }
 }
 
-pub trait ContractCallFields: TransactionCommonFields {
-    fn get_contract_account(&self) -> Result<AccountID> {
-        get_field(sfield::ContractAccount)
-    }
-
-    fn get_id(&self) -> Result<Hash256> {
-        get_field(sfield::ContractID)
-    }
-}
-
 #[cfg(test)]
 mod tests {
