+++ xrpl-wasm-stdlib/src/sfield.rs
@@ -161,7 +161,6 @@ pub const ExtensionSizeLimit: SField<u32, 131142> = SField::new();
 pub const GasPrice: SField<u32, 131143> = SField::new();
 pub const ComputationAllowance: SField<u32, 131144> = SField::new();
 pub const GasUsed: SField<u32, 131145> = SField::new();
-pub const ParameterFlag: SField<u32, 131146> = SField::new();
 pub const IndexNext: SField<u64, 196609> = SField::new();
 pub const IndexPrevious: SField<u64, 196610> = SField::new();
 pub const BookNode: SField<u64, 196611> = SField::new();
@@ -221,8 +220,6 @@ pub const VaultID: SField<Hash256, 327715> = SField::new();
 pub const ParentBatchID: SField<Hash256, 327716> = SField::new();
 pub const LoanBrokerID: SField<Hash256, 327717> = SField::new();
 pub const LoanID: SField<Hash256, 327718> = SField::new();
-pub const ContractHash: SField<Hash256, 327719> = SField::new();
-pub const ContractID: SField<Hash256, 327720> = SField::new();
 pub const Amount: SField<Amount, 393217> = SField::new();
 pub const Balance: SField<Amount, 393218> = SField::new();
 pub const LimitAmount: SField<Amount, 393219> = SField::new();
@@ -277,8 +274,6 @@ pub const Provider: SField<StandardBlob, 458781> = SField::new();
 pub const MPTokenMetadata: SField<StandardBlob, 458782> = SField::new();
 pub const CredentialType: SField<StandardBlob, 458783> = SField::new();
 pub const FinishFunction: SField<WasmBlob, 458784> = SField::new();
-pub const ContractCode: SField<StandardBlob, 458785> = SField::new();
-pub const FunctionName: SField<StandardBlob, 458786> = SField::new();
 pub const Account: SField<AccountID, 524289> = SField::new();
 pub const Owner: SField<AccountID, 524290> = SField::new();
 pub const Destination: SField<AccountID, 524291> = SField::new();
@@ -299,7 +294,6 @@ pub const IssuingChainDoor: SField<AccountID, 524311> = SField::new();
 pub const Subject: SField<AccountID, 524312> = SField::new();
 pub const Borrower: SField<AccountID, 524313> = SField::new();
 pub const Counterparty: SField<AccountID, 524314> = SField::new();
-pub const ContractAccount: SField<AccountID, 524315> = SField::new();
 pub const Number: SField<u8, 589825> = SField::new();
 pub const AssetsAvailable: SField<u8, 589826> = SField::new();
 pub const AssetsMaximum: SField<u8, 589827> = SField::new();
@@ -348,10 +342,6 @@ pub const RawTransaction: SField<Object, 917538> = SField::new();
 pub const BatchSigner: SField<Object, 917539> = SField::new();
 pub const Book: SField<Object, 917540> = SField::new();
 pub const CounterpartySignature: SField<Object, 917541> = SField::new();
-pub const Function: SField<Object, 917542> = SField::new();
-pub const InstanceParameter: SField<Object, 917543> = SField::new();
-pub const InstanceParameterValue: SField<Object, 917544> = SField::new();
-pub const Parameter: SField<Object, 917545> = SField::new();
 pub const Signers: SField<Array, 983043> = SField::new();
 pub const SignerEntries: SField<Array, 983044> = SField::new();
 pub const Template: SField<Array, 983045> = SField::new();
@@ -374,10 +364,6 @@ pub const AcceptedCredentials: SField<Array, 983068> = SField::new();
 pub const Permissions: SField<Array, 983069> = SField::new();
 pub const RawTransactions: SField<Array, 983070> = SField::new();
 pub const BatchSigners: SField<Array, 983071> = SField::new();
-pub const Functions: SField<Array, 983072> = SField::new();
-pub const InstanceParameters: SField<Array, 983073> = SField::new();
-pub const InstanceParameterValues: SField<Array, 983074> = SField::new();
-pub const Parameters: SField<Array, 983075> = SField::new();
 pub const CloseResolution: SField<u8, 1048577> = SField::new();
 pub const Method: SField<u8, 1048578> = SField::new();
 pub const TransactionResult: SField<u8, 1048579> = SField::new();
@@ -387,7 +373,6 @@ pub const TickSize: SField<u8, 1048592> = SField::new();
 pub const UNLModifyDisabling: SField<u8, 1048593> = SField::new();
 pub const WasLockingChainSend: SField<u8, 1048595> = SField::new();
 pub const WithdrawalPolicy: SField<u8, 1048596> = SField::new();
-pub const ContractResult: SField<u8, 1048597> = SField::new();
 pub const TakerPaysCurrency: SField<Hash160, 1114113> = SField::new();
 pub const TakerPaysIssuer: SField<Hash160, 1114114> = SField::new();
 pub const TakerGetsCurrency: SField<Hash160, 1114115> = SField::new();
@@ -407,9 +392,6 @@ pub const Asset2: SField<Issue, 1572868> = SField::new();
 pub const XChainBridge: SField<u8, 1638401> = SField::new();
 pub const BaseAsset: SField<Currency, 1703937> = SField::new();
 pub const QuoteAsset: SField<Currency, 1703938> = SField::new();
-pub const ParameterValue: SField<u8, 1769473> = SField::new();
-pub const ParameterType: SField<u8, 1835009> = SField::new();
-pub const ContractJson: SField<u8, 1900545> = SField::new();
 pub const Transaction: SField<u8, 655425793> = SField::new();
 pub const LedgerEntry: SField<u8, 655491329> = SField::new();
 pub const Validation: SField<u8, 655556865> = SField::new();
